Search on enter.
Simple fix to make it possible to search by pressing `Enter`.

Minor problem: if the input field is empty and you press enter, the `Please enter a query` message is not shown. This is a bit annoying but it seems like a current limitation of `streamlit`, I couldn't think of a nice way to get around it.